Fix Go 1.20 compatibility: move declarations to dedicated files
